+++ packages/app/src/cli/services/webhook/trigger.test.ts
@@ -3,6 +3,7 @@ import {getWebhookSample} from './request-sample.js'
 import {requestApiVersions} from './request-api-versions.js'
 import {triggerLocalWebhook} from './trigger-local-webhook.js'
 import {requestTopics} from './request-topics.js'
+import {findInEnv, findApiKey, requestAppInfo} from './find-app-info.js'
 import {
   collectAddressAndMethod,
   collectApiVersion,
@@ -10,9 +11,10 @@ import {
   collectTopic,
   WebhookTriggerFlags,
 } from '../../prompts/webhook/options-prompt.js'
-import * as output from '@shopify/cli-kit/node/output'
+import {outputInfo, outputSuccess, consoleError} from '@shopify/cli-kit/node/output'
 import {ensureAuthenticatedPartners} from '@shopify/cli-kit/node/session'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
+import {renderConfirmationPrompt} from '@shopify/cli-kit/node/ui'
 
 const aToken = 'A_TOKEN'
 const samplePayload = '{ "sampleField": "SampleValue" }'
@@ -24,12 +26,16 @@ const aPort = '1234'
 const aUrlPath = '/a/url/path'
 const anAddress = 'https://example.org'
 
+vi.mock('@shopify/cli-kit')
+vi.mock('@shopify/cli-kit/node/output')
 vi.mock('@shopify/cli-kit/node/session')
+vi.mock('@shopify/cli-kit/node/ui')
 vi.mock('../../prompts/webhook/options-prompt.js')
 vi.mock('./request-sample.js')
 vi.mock('./request-api-versions.js')
 vi.mock('./request-topics.js')
 vi.mock('./trigger-local-webhook.js')
+vi.mock('./find-app-info.js')
 
 const emptyJson = '{}'
 const successDirectResponse = {
@@ -66,14 +72,12 @@ describe('webhookTriggerService', () => {
 
     vi.mocked(getWebhookSample).mockResolvedValue(response)
 
-    const outputSpy = vi.spyOn(output, 'consoleError')
-
     // When
     await webhookTriggerService(sampleFlags())
 
     // Then
     expectCalls(aVersion)
-    expect(outputSpy).toHaveBeenCalledWith(`Request errors:\n  · Some error\n  · Another error`)
+    expect(consoleError).toHaveBeenCalledWith(`Request errors:\n  · Some error\n  · Another error`)
   })
 
   it('Safe notification in case of unexpected request errors', async () => {
@@ -91,14 +95,12 @@ describe('webhookTriggerService', () => {
 
     vi.mocked(getWebhookSample).mockResolvedValue(response)
 
-    const outputSpy = vi.spyOn(output, 'consoleError')
-
     // When
     await webhookTriggerService(sampleFlags())
 
     // Then
     expectCalls(aVersion)
-    expect(outputSpy).toHaveBeenCalledWith(`Request errors:\n${JSON.stringify(response.userErrors)}`)
+    expect(consoleError).toHaveBeenCalledWith(`Request errors:\n${JSON.stringify(response.userErrors)}`)
   })
 
   it('notifies about real delivery being sent', async () => {
@@ -108,8 +110,6 @@ describe('webhookTriggerService', () => {
     vi.mocked(triggerLocalWebhook)
     vi.mocked(getWebhookSample).mockResolvedValue(successEmptyResponse)
 
-    const outputSpy = vi.spyOn(output, 'outputSuccess')
-
     // When
     await webhookTriggerService(sampleFlags())
 
@@ -117,7 +117,7 @@ describe('webhookTriggerService', () => {
     expectCalls(aVersion)
     expect(getWebhookSample).toHaveBeenCalledWith(aToken, aTopic, aVersion, 'http', anAddress, aSecret)
     expect(triggerLocalWebhook).toHaveBeenCalledTimes(0)
-    expect(outputSpy).toHaveBeenCalledWith('Webhook has been enqueued for delivery')
+    expect(outputSuccess).toHaveBeenCalledWith('Webhook has been enqueued for delivery')
   })
 
   describe('Localhost delivery', () => {
@@ -128,8 +128,6 @@ describe('webhookTriggerService', () => {
       vi.mocked(triggerLocalWebhook).mockResolvedValue(true)
       vi.mocked(getWebhookSample).mockResolvedValue(successDirectResponse)
 
-      const outputSpy = vi.spyOn(output, 'outputSuccess')
-
       // When
       await webhookTriggerService(sampleFlags())
 
@@ -137,7 +135,7 @@ describe('webhookTriggerService', () => {
       expectCalls(aVersion)
       expect(getWebhookSample).toHaveBeenCalledWith(aToken, aTopic, aVersion, 'localhost', aFullLocalAddress, aSecret)
       expect(triggerLocalWebhook).toHaveBeenCalledWith(aFullLocalAddress, samplePayload, sampleHeaders)
-      expect(outputSpy).toHaveBeenCalledWith('Localhost delivery sucessful')
+      expect(outputSuccess).toHaveBeenCalledWith('Localhost delivery sucessful')
     })
 
     it('shows an error if localhost is not ready', async () => {
@@ -147,8 +145,6 @@ describe('webhookTriggerService', () => {
       vi.mocked(triggerLocalWebhook).mockResolvedValue(false)
       vi.mocked(getWebhookSample).mockResolvedValue(successDirectResponse)
 
-      const outputSpy = vi.spyOn(output, 'consoleError')
-
       // When
       await webhookTriggerService(sampleFlags())
 
@@ -156,7 +152,113 @@ describe('webhookTriggerService', () => {
       expectCalls(aVersion)
       expect(getWebhookSample).toHaveBeenCalledWith(aToken, aTopic, aVersion, 'localhost', aFullLocalAddress, aSecret)
       expect(triggerLocalWebhook).toHaveBeenCalledWith(aFullLocalAddress, samplePayload, sampleHeaders)
-      expect(outputSpy).toHaveBeenCalledWith('Localhost delivery failed')
+      expect(consoleError).toHaveBeenCalledWith('Localhost delivery failed')
+    })
+  })
+
+  describe('Shared secret choices', () => {
+    beforeEach(async () => {
+      vi.mocked(requestApiVersions).mockResolvedValue([aVersion])
+      vi.mocked(collectApiVersion).mockResolvedValue(aVersion)
+      vi.mocked(requestTopics).mockResolvedValue([aTopic])
+      vi.mocked(collectTopic).mockResolvedValue(aTopic)
+      vi.mocked(collectAddressAndMethod).mockResolvedValue(['localhost', aFullLocalAddress])
+
+      vi.mocked(triggerLocalWebhook).mockResolvedValue(true)
+      vi.mocked(getWebhookSample).mockResolvedValue(successDirectResponse)
+    })
+
+    it('manual', async () => {
+      // Given
+      vi.mocked(renderConfirmationPrompt).mockResolvedValue(false)
+      vi.mocked(collectSecret).mockResolvedValue(aSecret)
+
+      // When
+      await webhookTriggerService(noSecretSampleFlags())
+
+      // Then
+      expect(renderConfirmationPrompt).toHaveBeenCalledOnce()
+      expect(collectSecret).toHaveBeenCalledOnce()
+    })
+
+    it('uses .env credentials', async () => {
+      // Given
+      vi.mocked(renderConfirmationPrompt).mockResolvedValue(true)
+      vi.mocked(findInEnv).mockResolvedValue({clientSecret: aSecret})
+
+      // When
+      await webhookTriggerService(noSecretSampleFlags())
+
+      // Then
+      expect(renderConfirmationPrompt).toHaveBeenCalledOnce()
+      expect(collectSecret).toHaveBeenCalledTimes(0)
+      expect(requestAppInfo).toHaveBeenCalledTimes(0)
+      expect(outputInfo).toHaveBeenCalledWith('Reading client-secret from .env file')
+    })
+
+    it('no remote apiKey found', async () => {
+      // Given
+      vi.mocked(renderConfirmationPrompt).mockResolvedValue(true)
+      vi.mocked(findInEnv).mockResolvedValue({})
+      vi.mocked(findApiKey).mockResolvedValue(undefined)
+
+      // When
+      await webhookTriggerService(noSecretSampleFlags())
+
+      // Then
+      expect(renderConfirmationPrompt).toHaveBeenCalledOnce()
+      expect(collectSecret).toHaveBeenCalledOnce()
+      expect(requestAppInfo).toHaveBeenCalledTimes(0)
+    })
+
+    it('remote apiKey found', async () => {
+      // Given
+      vi.mocked(renderConfirmationPrompt).mockResolvedValue(true)
+      vi.mocked(findInEnv).mockResolvedValue({})
+      vi.mocked(findApiKey).mockResolvedValue('API_KEY')
+      vi.mocked(requestAppInfo).mockResolvedValue({clientSecret: aSecret, apiKey: 'API_KEY', clientId: 'Id'})
+
+      // When
+      await webhookTriggerService(noSecretSampleFlags())
+
+      // Then
+      expect(renderConfirmationPrompt).toHaveBeenCalledOnce()
+      expect(collectSecret).toHaveBeenCalledTimes(0)
+      expect(outputInfo).toHaveBeenCalledWith('Reading client-secret from app settings in Partners')
+      expect(requestAppInfo).toHaveBeenCalledWith(aToken, 'API_KEY')
+    })
+
+    it('remote apiKey found', async () => {
+      // Given
+      vi.mocked(renderConfirmationPrompt).mockResolvedValue(true)
+      vi.mocked(findInEnv).mockResolvedValue({})
+      vi.mocked(findApiKey).mockResolvedValue('API_KEY')
+      vi.mocked(requestAppInfo).mockResolvedValue({clientSecret: aSecret, apiKey: 'API_KEY', clientId: 'Id'})
+
+      // When
+      await webhookTriggerService(noSecretSampleFlags())
+
+      // Then
+      expect(renderConfirmationPrompt).toHaveBeenCalledOnce()
+      expect(collectSecret).toHaveBeenCalledTimes(0)
+      expect(outputInfo).toHaveBeenCalledWith('Reading client-secret from app settings in Partners')
+      expect(requestAppInfo).toHaveBeenCalledWith(aToken, 'API_KEY')
+    })
+
+    it('remote apiKey found but no app info', async () => {
+      // Given
+      vi.mocked(renderConfirmationPrompt).mockResolvedValue(true)
+      vi.mocked(findInEnv).mockResolvedValue({})
+      vi.mocked(findApiKey).mockResolvedValue('API_KEY')
+      vi.mocked(requestAppInfo).mockResolvedValue({})
+
+      // When
+      await webhookTriggerService(noSecretSampleFlags())
+
+      // Then
+      expect(renderConfirmationPrompt).toHaveBeenCalledOnce()
+      expect(collectSecret).toHaveBeenCalledOnce()
+      expect(requestAppInfo).toHaveBeenCalledWith(aToken, 'API_KEY')
     })
   })
 
@@ -179,8 +281,19 @@ describe('webhookTriggerService', () => {
       topic: aTopic,
       apiVersion: aVersion,
       deliveryMethod: 'event-bridge',
-      sharedSecret: aSecret,
+      clientSecret: aSecret,
       address: '',
+    }
+
+    return flags
+  }
+
+  function noSecretSampleFlags(): WebhookTriggerFlags {
+    const flags: WebhookTriggerFlags = {
+      topic: aTopic,
+      apiVersion: aVersion,
+      deliveryMethod: 'localhost',
+      address: aFullLocalAddress,
     }
 
     return flags