+++ packages/cli-kit/src/public/common/ts/deep-required.ts
@@ -1,5 +1,3 @@
-export {}
-
 /**
  * Converts a mapping type to be non-optional.
  *