+++ fixtures/app/CHANGELOG.md
@@ -1,5 +1,37 @@
 # app
 
+## 3.43.0
+
+### Patch Changes
+
+- Updated dependencies [4b0cc57ce]
+  - @shopify/app@3.43.0
+  - @shopify/cli@3.43.0
+
+## 3.42.0
+
+### Patch Changes
+
+- Updated dependencies [66653aecd]
+  - @shopify/app@3.42.0
+  - @shopify/cli@3.42.0
+
+## 3.41.2
+
+### Patch Changes
+
+- Updated dependencies [dd3d0c846]
+  - @shopify/app@3.41.2
+  - @shopify/cli@3.41.2
+
+## 3.41.1
+
+### Patch Changes
+
+- Updated dependencies [830617a4d]
+  - @shopify/app@3.41.1
+  - @shopify/cli@3.41.1
+
 ## 3.41.0
 
 ### Patch Changes