+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @shopify/create-app
 
+## 3.43.0
+
+### Patch Changes
+
+- Updated dependencies [f732207fa]
+- Updated dependencies [4b0cc57ce]
+- Updated dependencies [b6f93cfa7]
+  - @shopify/cli-kit@3.43.0
+
+## 3.42.0
+
+### Patch Changes
+
+- Updated dependencies [2203d4e6f]
+  - @shopify/cli-kit@3.42.0
+
+## 3.41.2
+
+### Patch Changes
+
+- @shopify/cli-kit@3.41.2
+
+## 3.41.1
+
+### Patch Changes
+
+- @shopify/cli-kit@3.41.1
+
 ## 3.41.0
 
 ### Patch Changes